check: Actually get input from current buffer (and fix test)
Previously, ledger-check-buffer was just getting its input from the default
journal file (e.g., configured by ~/.ledgerrc).

This has the unfortunate drawback of causing the errors reported by ledger to
come from a file called "", which is a little confusing.

Update ledger-report-visit-source to visit the marker's buffer rather than the
file name, which is now wrong.

This, in turn, forces us to fix the test which was mocking out shell-command,
which is no longer being used.  Instead, actually invoke the ledger binary.